Adding form to update values of json in friendly UI way.
