+++ app/renderer.js
@@ -1,7 +1,7 @@
 const { remote, shell } = require("electron");
-const Prism = require("prismjs");
 const flyd = require("flyd");
 const path = require("path");
+const { toHTML } = require("./json-views/index.js");
 
 const { ComponentApplicationMenu } = require("./components/application-menu/index.js");
 const { ComponentDragAndDrop } = require("./components/drag-and-drop/index.js");
@@ -9,6 +9,7 @@ const { ComponentSchemaButtons } = require("./components/schema-buttons/index.js
 const { ComponentSchemaErrors } = require("./components/schema-errors/index.js");
 const { ComponentTabs } = require("./components/tabs/index.js");
 const { FormUpdate } = require("./components/form-update/index.js");
+const { TextArea } = require("./components/text-area/index.js");
 const validator = require("./validator/index.js");
 
 const mainProcess = remote.require("./main.js");
@@ -24,6 +25,7 @@ const updateJSONStream = flyd.stream();
 const objPathStream = flyd.stream();
 const listJSONStream = flyd.stream();
 const schemaJSONStream = flyd.stream(null);
+const htmlStream = flyd.map(toHTML, inputJSONStream);
 
 const isSchemaValidStream = flyd.map((x) => validator.schema.validateSchema(schemaJSONStream(), x), inputJSONStream);
 const isJSONValidStream = flyd.map((x) => validator.input.validate(x), inputJSONStream);
@@ -32,22 +34,14 @@ ComponentDragAndDrop({ shell }, { inputJSONStream });
 ComponentApplicationMenu({ mainProcess, shell, currentWindow }, { applicationTitleStream, inputJSONStream, schemaJSONStream });
 ComponentDragAndDrop({ shell }, { inputJSONStream });
 ComponentSchemaButtons({}, { inputJSONStream, schemaJSONStream, isSchemaValidStream, isJSONValidStream });
-ComponentTabs({}, { inputJSONStream });
-ComponentTabs({}, { inputJSONStream });
+ComponentTabs({}, { inputJSONStream, htmlStream  });
 ComponentSchemaErrors({}, { schemaJSONStream, isSchemaValidStream });
-FormUpdate({}, { updateJSONStream, listJSONStream, inputJSONStream, objPathStream });
+FormUpdate({}, { updateJSONStream, listJSONStream, inputJSONStream, objPathStream, htmlStream });
+TextArea({}, { inputJSONStream });
 // Listeners
 inputJSON.addEventListener("input", (e) => inputJSONStream(e.target.value));
 listJSON.addEventListener("click", listJSONStream);
 updateJSON.addEventListener("click", updateJSONStream);
-
-flyd.on((value) => {
-    // Returns a highlighted HTML string
-    const html = value ? Prism.highlight(value, Prism.languages.javascript, "javascript") : "";
-    document.querySelector("#outputJSON").innerHTML = html;
-
-    Prism.highlightAll();
-}, inputJSONStream);
 
 flyd.on((data) => {
     inputJSON.value = data;
